Change items after feedback JF

+++ MusicService.Domain/Models/Track.cs
@@ -6,7 +6,7 @@
         public Album Album { get; set; }
         public long DurationMs { get; set; }
         public bool Explicit { get; set; }
-        public long TrackNumber { get; set; }
+        public int TrackNumber { get; set; }
         public int DiscNumber { get; set; }
     }
 }
